move constants to separate file
