+++ src/app/docs/(v2)/constants.ts
@@ -1,0 +1,14 @@
+export const versionOptions = [
+  { label: '1.0.x', value: '1.0.x' },
+  { label: '1.1.x', value: '1.1.x' },
+  { label: '1.2.x', value: '1.2.x' },
+  { label: '1.3.x', value: '1.3.x' },
+  { label: '2.x.x (latest)', value: '2.x.x' },
+];
+
+export const frameworkOptions = [
+  { label: 'Next.js', value: 'nextjs' },
+  { label: 'React Router', value: 'react-router' },
+  { label: 'Remix', value: 'remix' },
+  { label: 'React', value: 'react' },
+];
