+++ src/app/docs/(v2)/layout.tsx
@@ -1,26 +1,7 @@
 import VersionSelect from '@/components/docs/VersionSelect';
 import DocsSidebar from '@/components/docs/DocsSidebar';
-import { notFound } from 'next/navigation';
-import React, { Suspense } from 'react';
 import MobileSidebar from '@/components/docs/MobileSidebar';
 import FrameworkSelect from '@/components/docs/FrameworkSelect';
-import { Metadata } from 'next';
-import { ResolvingMetadata } from 'next';
-
-export const versionOptions = [
-  { label: '1.0.x', value: '1.0.x' },
-  { label: '1.1.x', value: '1.1.x' },
-  { label: '1.2.x', value: '1.2.x' },
-  { label: '1.3.x', value: '1.3.x' },
-  { label: '2.x.x (latest)', value: '2.x.x' },
-];
-
-export const frameworkOptions = [
-  { label: 'Next.js', value: 'nextjs' },
-  { label: 'React Router', value: 'react-router' },
-  { label: 'Remix', value: 'remix' },
-  { label: 'React', value: 'react' },
-];
 
 export default function layout({ children }: { children: React.ReactNode }) {
   return (
